CA-71 clean up home component and moved logic to users and searchbar components
