+++ client/src/components/Users.jsx
@@ -1,4 +1,25 @@
-const Users = ({ users, meta, onPageChange }) => {
+import { useEffect, useState } from 'react'
+import { useSelector, useDispatch } from 'react-redux'
+import { searchUsers } from '../features/users/usersAction.js'
+
+const Users = () => {
+  const dispatch = useDispatch()
+  const { users, meta } = useSelector(state => state.users)
+  const [searchQuery, setSearchQuery] = useState('')
+
+  const fetchUsers = async ({ search = '', page = 1 } = {}) => {
+    await dispatch(searchUsers({ search, page }))
+  }
+
+  const handlePageChange = async page => {
+    await fetchUsers({ search: searchQuery, page })
+  }
+
+  // Should remove this and don't display the results by default. It should be performed on search only
+  useEffect(() => {
+    fetchUsers()
+  }, [])
+
   return (
     <div>
       <div style={{ display: 'flex', flexDirection: 'column', gap: '10px' }}>
@@ -25,7 +46,7 @@ const Users = ({ users, meta, onPageChange }) => {
       {users.length > 0 && (
         <div style={{ marginTop: '20px' }}>
           <button
-            onClick={() => onPageChange(meta.page - 1)}
+            onClick={() => handlePageChange(meta.page - 1)}
             disabled={meta.page <= 1}
           >
             Previous
@@ -34,7 +55,7 @@ const Users = ({ users, meta, onPageChange }) => {
             Page {meta.page} of {meta.totalPages}{' '}
           </span>
           <button
-            onClick={() => onPageChange(meta.page + 1)}
+            onClick={() => handlePageChange(meta.page + 1)}
             disabled={meta.page >= meta.totalPages}
           >
             Next
